requirements.txt
pydantic
asyncio
streamlit
google-adk
python-dotenv
a2a
